// src/enums/DbEnum.ts
export enum DbEnum {
	Points = "Points",
	Nodes = "Nodes",
	Tracks = "Tracks",
	ANodes = "ANodes",
	Stations = "Stations",
	AerialWay = "Aerialway",
	PointsPiste = "PointsPiste",
	NodesPiste = "NodesPiste",
	TracksPiste = "TracksPiste",
	ANodesPiste = "ANodesPiste",
	StationsPiste = "StationsPiste",
	AerialwayPiste = "AerialwayPiste",
	TourismPoint = "TourismPoint",
	Configuration = "Configuration"
}

export const dbEnumList = [DbEnum.Points, DbEnum.Nodes, DbEnum.Tracks, DbEnum.ANodes, DbEnum.Stations, DbEnum.AerialWay, DbEnum.PointsPiste, DbEnum.NodesPiste, DbEnum.TracksPiste, DbEnum.ANodesPiste, DbEnum.StationsPiste, DbEnum.AerialwayPiste, DbEnum.TourismPoint, DbEnum.Configuration]