// src/enums/highway.enum.ts
export enum HighwayEnum {
    Motorway = 'motorway',
    Trunk = 'trunk',
    Primary = 'primary',
    Secondary = 'secondary',
    Tertiary = 'tertiary',
    Unclassified = 'unclassified',
    Residential = 'residential',
    MotorwayLink = 'motorway_link',
    TrunkLink = 'trunk_link',
    PrimaryLink = 'primary_link',
    SecondaryLink = 'secondary_link',
    TertiaryLink = 'tertiary_link',
    LivingStreet = 'living_street',
    Service = 'service',
    Pedestrian = 'pedestrian',
    Track = 'track',
    BusGuideway = 'bus_guideway',
    Escape = 'escape',
    Raceway = 'raceway',
    Road = 'road',
    Busway = 'busway',
    Footway = 'footway',
    Bridleway = 'bridleway',
    Steps = 'steps',
    Corridor = 'corridor',
    Path = 'path',
    Cicleway = 'circleway',
    Proposed = 'proposed',
    Construction = 'construction',
    BusStop = 'bus_stop',
    Crossing = 'crossing',
    Elevator = 'elevator',
    EmergencyBay = 'emergency_bay',
    EmergencyAccessPoint = 'emergency_access_point',
    GiveWay = 'give_way',
    Milestone = 'milestone',
    MiniRoundabout = 'mini_roundabout',
    MotorwayJunction = 'motorway_junction',
    PassingPlace = 'passing_place',
    Platform = 'platform',
    RestArea = 'rest_area',
    SpeedCamera = 'speed_camera',
    StreetLamp = 'street_lamp',
    Services = 'services',
    Stop = 'stop',
    TrafficMirror = 'traffic_mirror',
    TrafficSignals = 'traffic_signals',
    Trailhead = 'trailhead',
    TurningCircle = 'turning_circle',
    TurningLoop = 'turning_loop',
    TollGantry = 'tool_gantry',
    Gondola = 'gondola',
    PistaEasy = "easy",
    PistaNovice = "easy",
    PistaIntermediate = "intermediate",
    PistaAdvanced = "advanced",
    PistaExpert = "advanced",
    Aerialway = "aerialway"
  }