// src/enums/surface.enum.ts
export enum SurfaceEnum {
	Paved = 'paved',
	Asphalt = 'asphalt',
	Chipseal = 'chipseal',
	Concrate = 'concrate',
    ConcrateLanes = 'concrate:lanes',
    ConcratePlates = 'concrate:plates',
    PavingStones = 'paving_stones',
    Sett = 'sett',
    UnhewnCobblestone = 'unhewn_cobblestone',
    Cobblestone = 'cobblestone',
    Metal = 'metal',
    Wood = 'wood',
    SteppingStones = 'stepping_stones',
    UserDefined = 'user_defined',
    Unpaved = 'unpaved',
    Compacted = 'compacted',
    FineGravel = 'fine_gravel',
    Gravel = 'gravel',
    Rock = 'rock',
    Pebblestone = 'pebblestone',
    Groung = 'ground',
    Dirt = 'dirt',
    Earth = 'earth',
    Grass = 'grass',
    GrassPaver = 'grass_paver',
    Mud = 'mud',
    Sand = 'sand',
    Woodchips = 'woodchips',
    Snow = 'snow',
    Ice = 'ice',
    Salt = 'salt',
    Clay = 'clay',
    Tartan = 'tartan',
    ArtificialTurf = 'artificial_turf',
    Acrylic = 'acrylic',
    MetalGrid = 'metal_grid',
    Carpet = 'carpet',
    AerialWay = "aerialway"
}